+++ GameObject.hpp
@@ -7,22 +7,22 @@
 class GameObject
 {
 public:
-    GameObject(const char *texturesheet, int x, int y, int width, int height);
+    GameObject(const char *texturesheet, float x, float y, int width, int height);
     ~GameObject();
 
-    void update(int p, int q);
+    void update(float p, float q);
     void Render();
-    int getX() { return xpos; }
-    int getY() { return ypos; }
-    void setX(int x){xpos = x;}
-    void setY(int y){ypos = y;}
+    float getX() { return xpos; }
+    float getY() { return ypos; }
+    void setX(float x){xpos = x;}
+    void setY(float y){ypos = y;}
     int getHeight() { return h; }
     int getWidth() { return w; }
     void handleEvents();
 
 protected:
-    int xpos;
-    int ypos;
+    float xpos;
+    float ypos;
     int w;
     int h;
