+++ Player.hpp
@@ -5,16 +5,16 @@
 class Player : public GameObject
 {
 public:
-    Player(const char *texturesheet, int x, int y, int width, int height);
+    Player(const char *texturesheet, float x, float y, int width, int height);
     ~Player();
 
     int getVelX() { return vel_X; }
     int getVelY() { return vel_Y; }
     void handleEvents();
     void Animate();
-    bool isMoving() { return moving;}
-    int getDirection() { return dir;}
-    int getMove(){return move;}
+    bool isMoving() { return moving; }
+    int getDirection() { return dir; }
+    int getMove() { return move; }
 
 private:
     int vel_X;
